pyproject.toml
[project]
name = "materialeditpython50"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "ipykernel>=6.29.5",
    "keyboard>=0.13.5",
    "lxml-html-clean>=0.4.2",
    "numpy>=2.2.5",
    "opencv-python>=4.11.0.86",
    "passlib>=1.7.4",
    "peewee>=3.17.9",
    "photoshop-python-api>=0.24.0",
    "pillow>=11.2.1",
    "pyautogui>=0.9.54",
    "pygetwindow>=0.0.9",
    "pymongo>=4.12.0",
    "pyperclip>=1.9.0",
    "pypinyin>=0.54.0",
    "python-pptx>=1.0.2",
    "python-pptx-text-replacer>=0.0.6",
    "pyyaml>=6.0.2",
    "pyzbar>=0.1.9",
    "requests-html>=0.10.0",
    "selenium>=4.31.0",
    "tomorrow3>=1.2.2",
    "uvicorn>=0.34.2",
    "win10toast>=0.9",
]
[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple/"
